+++ duplicatesuricate/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.4.3'
+__version__ = '0.4.4'
